reading degrees from RTD?!

+++ zephyr-firmware/sigurd/app/src/can_com.c
@@ -24,7 +24,7 @@ void can_tx_cb(const struct device *device, int error, void *user_data) {
     }
 }
 
-int submit_can_pkt(uint8_t id, const void *packet, size_t length) {
+int submit_can_pkt(uint16_t id, const void *packet, size_t length) {
     struct can_frame frame = {
             .flags = 0,
             .id = id,
