+++ zephyr-firmware/sigurd/app/src/can_com.h
@@ -10,7 +10,7 @@ extern "C" {
 
 int init_can();
 
-int submit_can_pkt(uint8_t id, const void *packet, size_t length);
+int submit_can_pkt(uint16_t id, const void *packet, size_t length);
 
 int add_filter_can(can_rx_callback_t rx_callback, struct can_filter filter, void *can_user_data);
